Split out getDowngradePlanFromPurchase
